Fix Poly.taylor_at in graph-mode. Fix showcase.ipynb.
The important change is that `Poly.degs` is now a tuple of integers, not a tensor!
Thus the degrees of the polynomial must be known when compiling to graph-mode.
It would be logical now to revert `engine.poly_prod` to an earlier version where degs are integers and we do not iterate through a tensor. This would enable to remove `tf.function` decorator from `poly_prod`.